refactor structure 1 and 2
