+++ 1-get-random-avatar/src/index.css
@@ -1,0 +1,2 @@
+/* src/index.css */
+@import 'tailwindcss';
