--- get-random-avatar/package.json
@@ -1,31 +1,0 @@
-{
-  "dependencies": {
-    "@tailwindcss/vite": "^4.0.13",
-    "react": "^19.0.0",
-    "react-dom": "^19.0.0"
-  },
-  "devDependencies": {
-    "@eslint/js": "^9.21.0",
-    "@types/react": "^19.0.10",
-    "@types/react-dom": "^19.0.4",
-    "@vitejs/plugin-react": "^4.3.4",
-    "autoprefixer": "^10.4.21",
-    "eslint": "^9.21.0",
-    "eslint-plugin-react-hooks": "^5.1.0",
-    "eslint-plugin-react-refresh": "^0.4.19",
-    "globals": "^15.15.0",
-    "postcss": "^8.5.3",
-    "tailwindcss": "^4.0.13",
-    "vite": "^6.2.0"
-  },
-  "name": "get-random-avatar",
-  "private": true,
-  "scripts": {
-    "build": "vite build",
-    "dev": "vite",
-    "lint": "eslint .",
-    "preview": "vite preview"
-  },
-  "type": "module",
-  "version": "0.0.0"
-}
