--- get-random-avatar/src/config.js
@@ -1,2 +1,0 @@
-export const VITE_UNSPLASH_ACCESS_KEY = import.meta.env.VITE_UNSPLASH_ACCESS_KEY
-// 从环境变量中读取 Unsplash API 密钥，注意：Vite 要求环境变量必须以 VITE_ 开头
