--- get-random-avatar/src/index.css
@@ -1,2 +1,0 @@
-/* src/index.css */
-@import 'tailwindcss';
