--- get-random-avatar/vite.config.js
@@ -1,10 +1,0 @@
-import { defineConfig } from 'vite'
-import react from '@vitejs/plugin-react'
-import  tailwindcss from '@tailwindcss/vite'
-
-// https://vite.dev/config/
-export default defineConfig({
-  plugins: [
-    react(), 
-    tailwindcss(),],
-})
